+++ src/Ext/TechnoType/Body.h
@@ -399,6 +399,13 @@ public:
 		Valueable<bool> KeepTargetOnMove;
 		Valueable<Leptons> KeepTargetOnMove_ExtraDistance;
 
+		Nullable<bool> DirectionalArmor;
+		Nullable<float> DirectionalArmor_FrontMultiplier;
+		Nullable<float> DirectionalArmor_SideMultiplier;
+		Nullable<float> DirectionalArmor_BackMultiplier;
+		Nullable<float> DirectionalArmor_FrontField;
+		Nullable<float> DirectionalArmor_BackField;
+
 		Valueable<int> Power;
 
 		Valueable<int> ExtraTurretCount;
@@ -848,6 +855,13 @@ public:
 			, BunkerableAnyway { false }
 			, KeepTargetOnMove { false }
 			, KeepTargetOnMove_ExtraDistance { Leptons(0) }
+
+			, DirectionalArmor {}
+			, DirectionalArmor_FrontMultiplier {}
+			, DirectionalArmor_SideMultiplier {}
+			, DirectionalArmor_BackMultiplier {}
+			, DirectionalArmor_FrontField {}
+			, DirectionalArmor_BackField {}
 
 			, Power { }